backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
black==20.8b1
boto3==1.17.39
botocore==1.20.39
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-admin-interface==0.14.1
django-colorfield==0.3.2
django-contrib-comments==2.0.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-fluent-comments==2.1
django-smart-selects==1.5.9
django-storages==1.11.1
django-summernote==0.8.11.6
django-tag-parser==3.2
django-templated-mail==1.1.1
django-threadedcomments==1.2
django-vote==2.2.0
djangoflutterwave==0.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.0.1
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pylint==2.6.0
python-akismet==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
rave-python==1.2.7
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
votes==1.0.1
webencodings==0.5.1
wrapt==1.12.1
zipp==3.4.0
